requirements.txt
numpy
Pillow
schedule
smartcrop
python-twitter
